+++ Sources/CRAW/__craw.c
@@ -19,7 +19,7 @@ int __craw_get_system_errno() {
 #include <ntsecapi.h>
 #endif
 
-int __craw_get_entropy_bytes(uint8_t *out, const size_t len) {
+int __craw_get_entropy_bytes(uint8_t *_Nonnull out, const size_t len) {
 	if (len > 256) {
 		return -1; // limit the length to avoid large requests
 	}
@@ -73,18 +73,18 @@ int __craw_get_entropy_bytes(uint8_t *out, const size_t len) {
 	#endif
 }
 
-void __craw_secure_zero_bytes(uint8_t *ptr, size_t size) {
-    volatile uint8_t *volatile p = ptr;
-    while (size--) {
-        *p++ = 0;
-    }
+void __craw_secure_zero_bytes(uint8_t *_Nonnull ptr, size_t size) {
+	volatile uint8_t *volatile p = ptr;
+	while (size--) {
+		*p++ = 0;
+	}
 }
 
-uint64_t __craw_assert_secure_zero_bytes(const uint8_t *volatile ptr, size_t size) {
-    const volatile uint8_t *volatile p = (const volatile uint8_t *)ptr;
+uint64_t __craw_assert_secure_zero_bytes(const uint8_t *_Nonnull volatile ptr, size_t size) {
+	const volatile uint8_t *volatile p = (const volatile uint8_t *)ptr;
 	volatile uint64_t sum = 0;
-    for (size_t i = 0; i < size; i++) {
+	for (size_t i = 0; i < size; i++) {
 		sum |= p[i];
-    }
+	}
 	return sum;
 }