+++ Sources/CRAW/__craw.h
@@ -14,6 +14,7 @@
 #include <errno.h>
 #include <stdbool.h>
 #include "time.h"
+#include <sys/mman.h>
 
 /// @brief a function that returns the current system errno for the process.
 /// @return the system errno
@@ -21,12 +22,12 @@ int __craw_get_system_errno();
 
 /// @brief capture entropy from the system. maximum of 256 bytes.
 /// @return 0 on success, errno on failure.
-int __craw_get_entropy_bytes(uint8_t *buf, const size_t len);
+int __craw_get_entropy_bytes(uint8_t *_Nonnull buf, const size_t len);
 
 // secure zeroing function
-void __craw_secure_zero_bytes(uint8_t *ptr, size_t size);
+void __craw_secure_zero_bytes(uint8_t *_Nonnull ptr, size_t size);
 
 // used to ensure that memory has been zeroed
-uint64_t __craw_assert_secure_zero_bytes(const uint8_t *volatile ptr, size_t size);
+uint64_t __craw_assert_secure_zero_bytes(const uint8_t *_Nonnull volatile ptr, size_t size);
 
 #endif // __CRAW_H