+++ Sources/RAW_chachapoly/ChaChaPoly.swift
@@ -30,14 +30,29 @@ public struct Context {
 	private var ctx:__crawdog_chachapoly_ctx
 
 	public init?(key:UnsafeRawBufferPointer) {
+		var newContext = __crawdog_chachapoly_ctx()
 		switch key.count {
 			case MemoryLayout<Key16>.size:
-				self.init(key:key.baseAddress!.load(as: Key16.self))
+				fallthrough
 			case MemoryLayout<Key32>.size:
-				self.init(key:key.baseAddress!.load(as: Key32.self))
+				_ = __crawdog_chachapoly_init(&newContext, key.baseAddress!, Int32(key.count))
 			default:
 				return nil
 		}
+		self.ctx = newContext
+	}
+	
+	public init?(key:UnsafeBufferPointer<UInt8>) {
+		var newContext = __crawdog_chachapoly_ctx()
+		switch key.count {
+			case MemoryLayout<Key16>.size:
+				fallthrough
+			case MemoryLayout<Key32>.size:
+				_ = __crawdog_chachapoly_init(&newContext, key.baseAddress!, Int32(key.count))
+			default:
+				return nil
+		}
+		self.ctx = newContext
 	}
 
 	// pointer to 32 byte key
@@ -54,15 +69,6 @@ public struct Context {
 		self.ctx = newContext
 	}
 	
-	// 32 byte key
-	public init<K32>(key:borrowing K32) where K32:RAW_staticbuff, K32.RAW_staticbuff_storetype == Key32.RAW_staticbuff_storetype {
-		var newContext = __crawdog_chachapoly_ctx()
-		key.RAW_access_staticbuff {
-			_ = __crawdog_chachapoly_init(&newContext, $0, Int32(MemoryLayout<K32>.size))
-		}
-		self.ctx = newContext
-	}
-	
 	// 16 byte key
 	public init<K16>(key:borrowing K16) where K16:RAW_staticbuff, K16.RAW_staticbuff_storetype == Key16.RAW_staticbuff_storetype{
 		var newContext = __crawdog_chachapoly_ctx()
@@ -71,6 +77,16 @@ public struct Context {
 		}
 		self.ctx = newContext
 	}
+	
+	// 32 byte key
+	public init<K32>(key:borrowing K32) where K32:RAW_staticbuff, K32.RAW_staticbuff_storetype == Key32.RAW_staticbuff_storetype {
+		var newContext = __crawdog_chachapoly_ctx()
+		key.RAW_access_staticbuff {
+			_ = __crawdog_chachapoly_init(&newContext, $0, Int32(MemoryLayout<K32>.size))
+		}
+		self.ctx = newContext
+	}
+
 
 	/// execute authenticated encryption with associated data.
 	/// - parameters:
@@ -95,6 +111,18 @@ public struct Context {
 		}
 	}
 
+	public mutating func encrypt(nonce noncePtr:UnsafeRawPointer, associatedData:UnsafeRawBufferPointer, inputData:UnsafeRawBufferPointer, output:UnsafeMutableRawPointer, tag tagBuff:UnsafeMutableRawPointer) throws {
+		switch __crawdog_chachapoly_crypt(&ctx, noncePtr, associatedData.baseAddress, Int32(associatedData.count), inputData.baseAddress, Int32(inputData.count), output, tagBuff, Int32(MemoryLayout<Tag>.size), 1) {
+			case 0:
+				return
+			case __CRAWDOG_CHACHAPOLY_INVALID_MAC:
+				throw InvalidMACError()
+			default:
+				fatalError("unknown error thrown from rawdog chachapoly impl")
+		}
+	}
+
+
 	/// execute authenticated decryption with associated data.
 	/// - parameters:
 	///		- tag: the tag to authenticate the decryption
@@ -106,6 +134,17 @@ public struct Context {
 				__crawdog_chachapoly_crypt(&ctx, noncePtr, associatedData.baseAddress, Int32(associatedData.count), inputData.baseAddress, Int32(inputData.count), output, tagPtr, Int32(MemoryLayout<Tag>.size), 0)
 			}
 		}) {
+			case 0:
+				return
+			case __CRAWDOG_CHACHAPOLY_INVALID_MAC:
+				throw InvalidMACError()
+			default:
+				fatalError("unknown error thrown from rawdog chachapoly impl")
+		}
+	}
+
+	public mutating func decrypt(tag:UnsafeRawPointer, nonce:UnsafeRawPointer, associatedData:UnsafeRawBufferPointer, inputData:UnsafeRawBufferPointer, output:UnsafeMutableRawPointer) throws {
+		switch __crawdog_chachapoly_crypt(&ctx, nonce, associatedData.baseAddress, Int32(associatedData.count), inputData.baseAddress, Int32(inputData.count), output, UnsafeMutableRawPointer(mutating:tag), Int32(MemoryLayout<Tag>.size), 0) {
 			case 0:
 				return
 			case __CRAWDOG_CHACHAPOLY_INVALID_MAC: